# documentacao/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/simular-investimento":{"post":{"tags":["simulacao-controller"],"operationId":"simular","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimulacaoInvestimentoInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimulacaoInvestimentoOutput"}}}}}}},"/clientes":{"get":{"tags":["cliente-controller"],"operationId":"listar","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Cliente"}}}}}}},"post":{"tags":["cliente-controller"],"operationId":"criarCliente","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Cliente"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cliente"}}}}}}},"/auth/token":{"post":{"tags":["auth-controller"],"operationId":"generateToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/telemetria":{"get":{"tags":["telemetria-controller"],"operationId":"pesquisarDados","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RelatorioTelemetriaRecord"}}}}}}},"/simulacoes":{"get":{"tags":["simulacao-controller"],"operationId":"listarTodos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Simulacao"}}}}}}}},"/simulacoes/por-produto-dia/{dataSimulacao}":{"get":{"tags":["simulacao-controller"],"operationId":"buscarPorData","parameters":[{"name":"dataSimulacao","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Simulacao"}}}}}}}},"/produtos":{"get":{"tags":["produto-controller"],"operationId":"buscarTodosProdutos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Produto"}}}}}}}},"/produtos-recomendados/{idPerfil}":{"get":{"tags":["recomendacao-controller"],"operationId":"buscarPorIdPerfil","parameters":[{"name":"idPerfil","in":"path","required":true,"schema":{"type":"string","format":"byte"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Produto"}}}}}}}},"/perfis":{"get":{"tags":["perfil-controller"],"operationId":"buscarTodosPerfis","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Perfil"}}}}}}}},"/perfis/{id}":{"get":{"tags":["perfil-controller"],"operationId":"buscarPerfilPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"byte"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Perfil"}}}}}}},"/perfil-risco/{clienteId}":{"get":{"tags":["perfil-risco-controller"],"operationId":"buscarPerfil","parameters":[{"name":"clienteId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PerfilRiscoResponseRecord"}}}}}}},"/investimentos/{clienteId}":{"get":{"tags":["investimento-controller"],"operationId":"buscarPorCliente","parameters":[{"name":"clienteId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InvestimentoResponseRecord"}}}}}}}}},"components":{"schemas":{"SimulacaoInvestimentoInput":{"required":["clienteId","prazoMeses","tipoProduto","valor"],"type":"object","properties":{"clienteId":{"type":"integer","format":"int64"},"valor":{"type":"number"},"prazoMeses":{"type":"integer","format":"int32"},"tipoProduto":{"maxLength":2147483647,"minLength":1,"type":"string"}}},"ProdutoValidado":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"tipo":{"type":"string"},"rentabilidade":{"type":"number"},"risco":{"type":"string"}}},"ResultadoRetornoSimulacao":{"type":"object","properties":{"valorFinal":{"type":"number"},"rentabilidadeEfetiva":{"type":"number"},"prazoMeses":{"type":"integer","format":"int32"}}},"SimulacaoInvestimentoOutput":{"type":"object","properties":{"produtoValidado":{"$ref":"#/components/schemas/ProdutoValidado"},"resultadoSimulacao":{"$ref":"#/components/schemas/ResultadoRetornoSimulacao"},"dataSimulacao":{"type":"string","format":"date-time"}}},"Cliente":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"}}},"AuthRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"}}},"PeriodoRecord":{"type":"object","properties":{"inicio":{"type":"string","format":"date"},"fim":{"type":"string","format":"date"}}},"RelatorioTelemetriaRecord":{"type":"object","properties":{"servicos":{"type":"array","items":{"$ref":"#/components/schemas/ServicoTelemetriaRecord"}},"periodo":{"$ref":"#/components/schemas/PeriodoRecord"}}},"ServicoTelemetriaRecord":{"type":"object","properties":{"nome":{"type":"string"},"quantidadeChamadas":{"type":"integer","format":"int64"},"mediaTempoRespostaMs":{"type":"number","format":"double"}}},"Produto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"tipo":{"type":"string"},"rentabilidade":{"type":"number"},"risco":{"type":"string"},"classificacaoLiquidez":{"type":"integer","format":"int32"}}},"Simulacao":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cliente":{"$ref":"#/components/schemas/Cliente"},"produto":{"$ref":"#/components/schemas/Produto"},"vlInvestido":{"type":"number"},"vlFinal":{"type":"number"},"qtPrazo":{"type":"integer","format":"int32"},"dtSimulacao":{"type":"string","format":"date"}}},"Perfil":{"type":"object","properties":{"id":{"type":"string","format":"byte"},"nome":{"type":"string"},"descricao":{"type":"string"}}},"PerfilRiscoResponseRecord":{"type":"object","properties":{"clienteId":{"type":"integer","format":"int64"},"perfil":{"type":"string"},"pontuacao":{"type":"integer","format":"int32"},"descricao":{"type":"string"}}},"InvestimentoResponseRecord":{"type":"object","properties":{"idInvestimento":{"type":"integer","format":"int64"},"nomeProduto":{"type":"string"},"tipoProduto":{"type":"string"},"valorInvestimento":{"type":"number"},"dataInvestimento":{"type":"string","format":"date-time"}}}}}}